rename to more generic reducer and provider names
