+++ app/javascript/components/supplemental_application/SupplementalApplicationPage.js
@@ -8,7 +8,7 @@ import Alerts from 'components/Alerts'
 import Loading from 'components/molecules/Loading'
 import LeaveConfirmationModal from 'components/organisms/LeaveConfirmationModal'
 import { getPageHeaderData } from 'components/supplemental_application/leaseUpApplicationBreadcrumbs'
-import { AppContext } from 'context/LeaseUpProvider'
+import { AppContext } from 'context/Provider'
 import appPaths from 'utils/appPaths'
 import { useEffectOnMount, useStateObject } from 'utils/customHooks'
 import formUtils from 'utils/formUtils'
