+++ app/javascript/routes/LeaseUpApp.js
@@ -2,16 +2,16 @@ import React from 'react'
 
 import { BrowserRouter as Router } from 'react-router-dom'
 
-import LeaseUpProvider from 'context/LeaseUpProvider'
+import Provider from 'context/Provider'
 
 import LeaseUpRoutes from './LeaseUpRoutes'
 
 const LeaseUpApp = () => (
-  <LeaseUpProvider>
+  <Provider>
     <Router>
       <LeaseUpRoutes />
     </Router>
-  </LeaseUpProvider>
+  </Provider>
 )
 
 export default LeaseUpApp
